% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afnireg-package.R
\docType{package}
\name{afnireg-package}
\alias{afnireg}
\alias{afnireg-package}
\title{afnireg: AFNI Integration for fMRI Analysis in R}
\description{
The afnireg package provides an R interface to AFNI's 3dDeconvolve
program for fMRI analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/afnireg}
  \item Report bugs at \url{https://github.com/bbuchsbaum/afnireg/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
