% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_generics.R
\name{build_afni_stims}
\alias{build_afni_stims}
\title{Build AFNI Stimulus Files}
\usage{
build_afni_stims(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Additional arguments}
}
\description{
Generic function to build AFNI stimulus files
}
