% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_s3_methods.R
\name{cells.afni_hrf_convolved_term}
\alias{cells.afni_hrf_convolved_term}
\title{Get cells for afni_hrf_convolved_term}
\usage{
\method{cells}{afni_hrf_convolved_term}(x, ...)
}
\arguments{
\item{x}{An afni_hrf_convolved_term object}

\item{...}{Additional arguments}
}
\description{
Get cells for afni_hrf_convolved_term
}
