% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_construct.R
\name{construct.afni_trialwise_hrfspec}
\alias{construct.afni_trialwise_hrfspec}
\title{Construct an afni_trialwise_convolved_term from an afni_trialwise_hrfspec}
\usage{
\method{construct}{afni_trialwise_hrfspec}(x, model_spec, ...)
}
\arguments{
\item{x}{An afni_trialwise_hrfspec object}

\item{model_spec}{Model specification containing data and sampling frame}

\item{...}{Additional arguments}
}
\description{
Construct an afni_trialwise_convolved_term from an afni_trialwise_hrfspec
}
