% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_contrasts.R
\name{contrast_weights.afni_hrf_convolved_term}
\alias{contrast_weights.afni_hrf_convolved_term}
\title{Extract contrast weights for an afni_hrf_convolved_term}
\usage{
\method{contrast_weights}{afni_hrf_convolved_term}(x, ...)
}
\description{
Extract contrast weights for an afni_hrf_convolved_term
}
\keyword{internal}
