% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{design_matrix.afni_hrf_convolved_term}
\alias{design_matrix.afni_hrf_convolved_term}
\title{Extract Design Matrix from AFNI Convolved Terms}
\usage{
\method{design_matrix}{afni_hrf_convolved_term}(x, ...)
}
\arguments{
\item{x}{An AFNI convolved term object}

\item{...}{Additional arguments}
}
\value{
NULL or empty matrix
}
\description{
AFNI convolved terms don't produce R design matrices as they are
processed externally by 3dDeconvolve.
}
