% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{is_afni_convolved_term}
\alias{is_afni_convolved_term}
\title{Check if an Object is an AFNI Convolved Term}
\usage{
is_afni_convolved_term(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical indicating if the object is an AFNI convolved term
}
\description{
Check if an Object is an AFNI Convolved Term
}
