% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{nbasis.afni_hrf_convolved_term}
\alias{nbasis.afni_hrf_convolved_term}
\title{Get Number of Basis Functions for AFNI Terms}
\usage{
\method{nbasis}{afni_hrf_convolved_term}(x, ...)
}
\arguments{
\item{x}{An AFNI convolved term object}

\item{...}{Additional arguments (not used)}
}
\value{
Number of basis functions
}
\description{
Get Number of Basis Functions for AFNI Terms
}
