% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{nbasis.event_term}
\alias{nbasis.event_term}
\title{Get Number of Basis Functions for Event Terms}
\usage{
\method{nbasis}{event_term}(x, ...)
}
\arguments{
\item{x}{An event_term object}

\item{...}{Additional arguments (not used)}
}
\value{
Number of basis functions
}
\description{
Get Number of Basis Functions for Event Terms
}
