% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_lm.R
\name{print.afni_lm_spec}
\alias{print.afni_lm_spec}
\title{Print method for afni_lm_spec}
\usage{
\method{print}{afni_lm_spec}(x, ...)
}
\arguments{
\item{x}{An afni_lm_spec object}

\item{...}{Additional arguments}
}
\description{
Print method for afni_lm_spec
}
