% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{print.afni_hrf_convolved_term}
\alias{print.afni_hrf_convolved_term}
\alias{print.afni_trialwise_convolved_term}
\title{AFNI-specific Convolved Term Classes}
\usage{
\method{print}{afni_hrf_convolved_term}(x, ...)

\method{print}{afni_trialwise_convolved_term}(x, ...)
}
\arguments{
\item{x}{An afni_trialwise_convolved_term object.}

\item{...}{Additional arguments.}
}
\description{
This file contains the AFNI-specific convolved term classes that were
previously defined in fmridesign. These classes represent terms that
are processed by AFNI's 3dDeconvolve rather than R's convolution.
}
\details{
Print method for afni_hrf_convolved_term objects.
}
