% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afnireg-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{event_model}
\alias{baseline_model}
\alias{event_term}
\alias{construct}
\alias{event_table}
\alias{split_onsets}
\alias{blocklens}
\alias{blockids}
\alias{sampling_frame}
\alias{fmri_dataset}
\alias{fmri_model}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fmridataset}{\code{\link[fmridataset]{fmri_dataset}}}

  \item{fmridesign}{\code{\link[fmridesign]{baseline_model}}, \code{\link[fmridesign]{blockids}}, \code{\link[fmridesign]{construct}}, \code{\link[fmridesign]{event_model}}, \code{\link[fmridesign]{event_table}}, \code{\link[fmridesign]{event_term}}, \code{\link[fmridesign]{split_onsets}}}

  \item{fmrihrf}{\code{\link[fmrihrf]{blocklens}}, \code{\link[fmrihrf]{sampling_frame}}}

  \item{fmrireg}{\code{\link[fmrireg]{fmri_model}}}
}}

