% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_classes.R
\name{requires_external_processing}
\alias{requires_external_processing}
\title{Check if a Term Requires External Processing}
\usage{
requires_external_processing(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical indicating if external processing is required
}
\description{
Check if a Term Requires External Processing
}
