% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_lm.R
\name{run.afni_lm_spec}
\alias{run.afni_lm_spec}
\title{Run AFNI Linear Model Specification}
\usage{
\method{run}{afni_lm_spec}(x, outdir, execute = TRUE, execfun = system, prepend = "", ...)
}
\arguments{
\item{x}{An afni_lm_spec object}

\item{outdir}{Output directory}

\item{execute}{Logical; execute the command}

\item{execfun}{Function to execute the command (default: system)}

\item{prepend}{String to prepend to command}

\item{...}{Additional arguments}
}
\value{
Result of execution
}
\description{
Execute an AFNI linear model specification using 3dDeconvolve.
}
