% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_stim_files.R
\name{write_afni_stim}
\alias{write_afni_stim}
\alias{write_afni_stim.afni_stim_file}
\alias{write_afni_stim.afni_stim_times}
\alias{write_afni_stim.afni_stim_im_times}
\title{Write AFNI stimulus files}
\usage{
write_afni_stim(x, ...)

\method{write_afni_stim}{afni_stim_file}(stim, dir)

\method{write_afni_stim}{afni_stim_times}(stim, dir)

\method{write_afni_stim}{afni_stim_im_times}(stim, dir)
}
\arguments{
\item{x}{An AFNI stimulus object}

\item{...}{Additional arguments (e.g., target directory)}

\item{stim}{For methods: stimulus object to write}

\item{dir}{For methods: output directory path}
}
\description{
Generic for writing AFNI stimulus specifications (values, times, or
individually-modulated times) to disk in AFNI-compatible formats.
}
