% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_build_stims.R, R/afni_generics.R
\name{build_afni_stims.convolved_term}
\alias{build_afni_stims.convolved_term}
\alias{build_afni_stims.afni_hrf_convolved_term}
\alias{build_afni_stims.afni_trialwise_convolved_term}
\alias{build_afni_stims.event_term}
\alias{build_afni_stims}
\title{Build AFNI Stimulus Files}
\usage{
\method{build_afni_stims}{convolved_term}(x, iresp = FALSE, tr_times = 1, ...)

\method{build_afni_stims}{afni_hrf_convolved_term}(x, iresp = FALSE, tr_times = 1, ...)

\method{build_afni_stims}{afni_trialwise_convolved_term}(x, iresp = FALSE, tr_times = 1, ...)

\method{build_afni_stims}{event_term}(x, iresp = FALSE, tr_times = 1, ...)

build_afni_stims(x, ...)
}
\arguments{
\item{x}{The input object (e.g., term)}

\item{iresp}{Logical; request impulse response outputs when supported}

\item{tr_times}{Numeric TR in seconds passed to timing-based terms}

\item{...}{Additional arguments, e.g., `iresp`, `tr_times`}
}
\description{
Generic function to build AFNI stimulus files from model terms.
Methods may accept additional arguments such as `iresp` (logical,
request impulse response outputs) and `tr_times` (numeric, TR in
seconds) used by AFNI-native HRF terms.
}
